bank dataset
The data is related to the direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed by the customer or not.
Banking_data.csv: 45,211 rows and 18 columns ordered by date (from May 2008 to November 2010)